refactor: Move chart-specific functions to charts dir
