+++ app/domain/index.ts
@@ -1,3 +1,2 @@
 export * from "./data";
-export * from "./charts";
 export * from "./chart-id";
